services/retrieval/pyproject.toml
[project]
name = "retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "httpx==0.28.1",
    "marshmallow<4.0.0",
    "pika==1.3.2",
    "psycopg2-binary==2.9.10",
    "pymilvus>=2.4.0",
    "python-dotenv==1.0.1",
    "sentence-transformers==3.3.1",
    "setuptools<70.0.0",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.34.0",
]
